requirements.txt
selene==2.0.0b16
pytest==7.2.0
allure-pytest
